Inkscape uses absolute paths in svg->pdf.
A known limitation of X11 in MacOSX results in a bug where Inkscape
fails when converting from SVG to PDF. For details of the bug itself,
see https://bugs.launchpad.net/inkscape/+bug/1449251. A fix is to use
absolute paths when specifying input and output files to Inkscape which
is implemented in the file bin/buildfigpdfs.